+++ source/scripting_v3/GTA/Entities/Peds/NavMeshBlockingObject.cs
@@ -35,7 +35,16 @@ namespace GTA
 
 			// Set the second last parameter bPermanent to false, which determines whether the blocking object will last outside the lifetime of the calling script (scrThread)
 			// If the SHVDN runtime stops working, all the blocking objects created via SHVDN will get deleted (stopping any SHVDN scripts working will not automatically remove any blocking objects)
-			int handle = Function.Call<int>(Hash.ADD_NAVMESH_BLOCKING_OBJECT, position.X, position.Y, position.Z, size.X, size.Y, size.Z, headingRadians, false, flags);
+			int handle = Function.Call<int>(Hash.ADD_NAVMESH_BLOCKING_OBJECT,
+				position.X,
+				position.Y,
+				position.Z,
+				size.X,
+				size.Y,
+				size.Z,
+				headingRadians,
+				false,
+				(int)flags);
 			return handle != -1 ? new NavMeshBlockingObject(handle) : null;
 		}
